+++ Source/Krypton Toolkit Examples/KryptonDateTimePicker Examples/Form1.Designer.cs
@@ -91,7 +91,7 @@
             this.rbOffice2010Black.TabIndex = 2;
             this.rbOffice2010Black.Text = "Office 2010 - Black";
             this.rbOffice2010Black.UseVisualStyleBackColor = true;
-            //this.rbOffice2010Black.CheckedChanged += new System.EventHandler(this.rbOffice2010Black_CheckedChanged);
+            this.rbOffice2010Black.CheckedChanged += new System.EventHandler(this.rbOffice2010Black_CheckedChanged);
             // 
             // rbOffice2010Silver
             // 
@@ -102,7 +102,7 @@
             this.rbOffice2010Silver.TabIndex = 1;
             this.rbOffice2010Silver.Text = "Office 2010 - Silver";
             this.rbOffice2010Silver.UseVisualStyleBackColor = true;
-            //this.rbOffice2010Silver.CheckedChanged += new System.EventHandler(this.rbOffice2010Silver_CheckedChanged);
+            this.rbOffice2010Silver.CheckedChanged += new System.EventHandler(this.rbOffice2010Silver_CheckedChanged);
             // 
             // rbOffice2010Blue
             // 
@@ -115,7 +115,7 @@
             this.rbOffice2010Blue.TabStop = true;
             this.rbOffice2010Blue.Text = "Office 2010 - Blue";
             this.rbOffice2010Blue.UseVisualStyleBackColor = true;
-            //this.rbOffice2010Blue.CheckedChanged += new System.EventHandler(this.rbOffice2010Blue_CheckedChanged);
+            this.rbOffice2010Blue.CheckedChanged += new System.EventHandler(this.rbOffice2010Blue_CheckedChanged);
             // 
             // rbSparklePurple
             // 
@@ -126,7 +126,7 @@
             this.rbSparklePurple.TabIndex = 8;
             this.rbSparklePurple.Text = "Sparkle - Purple";
             this.rbSparklePurple.UseVisualStyleBackColor = true;
-            //this.rbSparklePurple.CheckedChanged += new System.EventHandler(this.rbSparklePurple_CheckedChanged);
+            this.rbSparklePurple.CheckedChanged += new System.EventHandler(this.rbSparklePurple_CheckedChanged);
             // 
             // rbSparkleOrange
             // 
@@ -137,7 +137,7 @@
             this.rbSparkleOrange.TabIndex = 7;
             this.rbSparkleOrange.Text = "Sparkle - Orange";
             this.rbSparkleOrange.UseVisualStyleBackColor = true;
-            //this.rbSparkleOrange.CheckedChanged += new System.EventHandler(this.rbSparkleOrange_CheckedChanged);
+            this.rbSparkleOrange.CheckedChanged += new System.EventHandler(this.rbSparkleOrange_CheckedChanged);
             // 
             // rbSparkleBlue
             // 
@@ -148,7 +148,7 @@
             this.rbSparkleBlue.TabIndex = 6;
             this.rbSparkleBlue.Text = "Sparkle - Blue";
             this.rbSparkleBlue.UseVisualStyleBackColor = true;
-            //this.rbSparkleBlue.CheckedChanged += new System.EventHandler(this.rbSparkleBlue_CheckedChanged);
+            this.rbSparkleBlue.CheckedChanged += new System.EventHandler(this.rbSparkleBlue_CheckedChanged);
             // 
             // rbSystem
             // 
@@ -159,7 +159,7 @@
             this.rbSystem.TabIndex = 10;
             this.rbSystem.Text = "System";
             this.rbSystem.UseVisualStyleBackColor = true;
-            //this.rbSystem.CheckedChanged += new System.EventHandler(this.rbSystem_CheckedChanged);
+            this.rbSystem.CheckedChanged += new System.EventHandler(this.rbSystem_CheckedChanged);
             // 
             // rbOffice2003
             // 
@@ -170,7 +170,7 @@
             this.rbOffice2003.TabIndex = 9;
             this.rbOffice2003.Text = "Office 2003";
             this.rbOffice2003.UseVisualStyleBackColor = true;
-            //this.rbOffice2003.CheckedChanged += new System.EventHandler(this.rbOffice2003_CheckedChanged);
+            this.rbOffice2003.CheckedChanged += new System.EventHandler(this.rbOffice2003_CheckedChanged);
             // 
             // rbOffice2007Black
             // 
@@ -181,7 +181,7 @@
             this.rbOffice2007Black.TabIndex = 5;
             this.rbOffice2007Black.Text = "Office 2007 - Black";
             this.rbOffice2007Black.UseVisualStyleBackColor = true;
-            //this.rbOffice2007Black.CheckedChanged += new System.EventHandler(this.rbOffice2007Black_CheckedChanged);
+            this.rbOffice2007Black.CheckedChanged += new System.EventHandler(this.rbOffice2007Black_CheckedChanged);
             // 
             // rbOffice2007Silver
             // 
@@ -192,7 +192,7 @@
             this.rbOffice2007Silver.TabIndex = 4;
             this.rbOffice2007Silver.Text = "Office 2007 - Silver";
             this.rbOffice2007Silver.UseVisualStyleBackColor = true;
-            //this.rbOffice2007Silver.CheckedChanged += new System.EventHandler(this.rbOffice2007Silver_CheckedChanged);
+            this.rbOffice2007Silver.CheckedChanged += new System.EventHandler(this.rbOffice2007Silver_CheckedChanged);
             // 
             // rbOffice2007Blue
             // 
@@ -203,7 +203,7 @@
             this.rbOffice2007Blue.TabIndex = 3;
             this.rbOffice2007Blue.Text = "Office 2007 - Blue";
             this.rbOffice2007Blue.UseVisualStyleBackColor = true;
-            //this.rbOffice2007Blue.CheckedChanged += new System.EventHandler(this.rbOffice2007Blue_CheckedChanged);
+            this.rbOffice2007Blue.CheckedChanged += new System.EventHandler(this.rbOffice2007Blue_CheckedChanged);
             // 
             // groupBox4
             //