+++ Source/Krypton Toolkit Examples/KryptonDomainUpDown Examples/Form1.Designer.cs
@@ -509,7 +509,6 @@
             // kryptonPalette
             // 
             this.kryptonPalette.BaseFont = new System.Drawing.Font("Segoe UI", 9F);
-            this.kryptonPalette.BasePaletteType = Krypton.Toolkit.BasePaletteType.Microsoft365;
             this.kryptonPalette.BasePaletteType = Krypton.Toolkit.BasePaletteType.Custom;
             this.kryptonPalette.GridStyles.GridCustom1.StateCommon.Background.Color1 = System.Drawing.Color.FromArgb(((int)(((byte)(224)))), ((int)(((byte)(224)))), ((int)(((byte)(224)))));
             this.kryptonPalette.GridStyles.GridCustom1.StateCommon.Background.ColorStyle = Krypton.Toolkit.PaletteColorStyle.Solid;